gdb.base/break.exp C++ify
Some adjustments to make gdb.base/break.exp work when compiled as a
C++ program.  Passes cleanly with Clang++, but not with G++.  The
latter puts a breakpoint at an unexpected line in one case.  It seems
like a bug that gcc and g++ behave differently here.

gdb/testsuite/ChangeLog:

	* gdb.base/break.exp (func): New.  Use it throughout when
	expecting a function name.